Revert "Introduced event metadata"
This reverts commit f92092d326f7d1111e066e6ab45dba1452a3b084.
