+++ Source/EventFlow.Owin/MetadataProviders/AddRequestHeadersMetadataProvider.cs
@@ -49,7 +49,7 @@ namespace EventFlow.Owin.MetadataProviders
         public IEnumerable<KeyValuePair<string, string>> ProvideMetadata<TAggregate, TIdentity>(
             TIdentity id,
             IAggregateEvent aggregateEvent,
-            IEventMetadata eventMetadata)
+            IMetadata metadata)
             where TAggregate : IAggregateRoot<TIdentity>
             where TIdentity : IIdentity
         {
