+++ Source/EventFlow.Owin/MetadataProviders/AddUriMetadataProvider.cs
@@ -41,7 +41,7 @@ namespace EventFlow.Owin.MetadataProviders
         public IEnumerable<KeyValuePair<string, string>> ProvideMetadata<TAggregate, TIdentity>(
             TIdentity id,
             IAggregateEvent aggregateEvent,
-            IEventMetadata eventMetadata)
+            IMetadata metadata)
             where TAggregate : IAggregateRoot<TIdentity>
             where TIdentity : IIdentity
         {
