+++ Source/EventFlow.Owin/MetadataProviders/AddUserHostAddressMetadataProvider.cs
@@ -50,7 +50,7 @@ namespace EventFlow.Owin.MetadataProviders
         public IEnumerable<KeyValuePair<string, string>> ProvideMetadata<TAggregate, TIdentity>(
             TIdentity id,
             IAggregateEvent aggregateEvent,
-            IEventMetadata eventMetadata)
+            IMetadata metadata)
             where TAggregate : IAggregateRoot<TIdentity>
             where TIdentity : IIdentity
         {
