+++ Source/EventFlow.TestHelpers/Test.cs
@@ -73,7 +73,7 @@ namespace EventFlow.TestHelpers
             int aggregateSequenceNumber = 0)
             where TAggregateEvent : IAggregateEvent
         {
-            var metadata = new EventMetadata
+            var metadata = new Metadata
                 {
                     Timestamp = A<DateTimeOffset>(),
                     SourceId = A<SourceId>(),
