+++ Source/EventFlow.Tests/UnitTests/Aggregates/MetadataTests.cs
@@ -23,7 +23,6 @@
 using System;
 using System.Collections.Generic;
 using EventFlow.Aggregates;
-using EventFlow.Core;
 using EventFlow.TestHelpers;
 using FluentAssertions;
 using Newtonsoft.Json;
@@ -40,8 +39,8 @@ namespace EventFlow.Tests.UnitTests.Aggregates
             var timestamp = A<DateTimeOffset>();
 
             // Act
-            var sut = new EventMetadata
-            {
+            var sut = new Metadata
+                {
                     Timestamp = timestamp
                 };
 
@@ -56,8 +55,8 @@ namespace EventFlow.Tests.UnitTests.Aggregates
             var eventName = A<string>();
 
             // Act
-            var sut = new EventMetadata
-            {
+            var sut = new Metadata
+                {
                     EventName = eventName
                 };
 
@@ -72,8 +71,8 @@ namespace EventFlow.Tests.UnitTests.Aggregates
             var eventVersion = A<int>();
 
             // Act
-            var sut = new EventMetadata
-            {
+            var sut = new Metadata
+                {
                     EventVersion = eventVersion
                 };
 
@@ -88,7 +87,7 @@ namespace EventFlow.Tests.UnitTests.Aggregates
             var aggregateSequenceNumber = A<int>();
 
             // Act
-            var sut = new EventMetadata
+            var sut = new Metadata
                 {
                     AggregateSequenceNumber = aggregateSequenceNumber
                 };
@@ -126,7 +125,7 @@ namespace EventFlow.Tests.UnitTests.Aggregates
             var aggregateName = A<string>();
             var aggregateSequenceNumber = A<int>();
             var timestamp = A<DateTimeOffset>();
-            var sut = new EventMetadata
+            var sut = new Metadata
                 {
                     { MetadataKeys.AggregateName, aggregateName },
                     { MetadataKeys.AggregateSequenceNumber, aggregateSequenceNumber.ToString() },
@@ -135,7 +134,7 @@ namespace EventFlow.Tests.UnitTests.Aggregates
 
             // Act
             var json = JsonConvert.SerializeObject(sut);
-            var metadata = JsonConvert.DeserializeObject<EventMetadata>(json);
+            var metadata = JsonConvert.DeserializeObject<Metadata>(json);
 
             // Assert
             metadata.Count.Should().Be(3);
