+++ Source/EventFlow/Aggregates/Metadata.cs
@@ -22,14 +22,33 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using EventFlow.Core;
+using EventFlow.Exceptions;
 using EventFlow.Extensions;
 using Newtonsoft.Json;
 
 namespace EventFlow.Aggregates
 {
-    public class EventMetadata : Metadata, IEventMetadata
+    public class Metadata : Dictionary<string, string>, IMetadata
     {
+        public static IMetadata Empty { get; } = new Metadata();
+
+        public static IMetadata With(IEnumerable<KeyValuePair<string, string>> keyValuePairs)
+        {
+            return new Metadata(keyValuePairs);
+        }
+
+        public static IMetadata With(params KeyValuePair<string, string>[] keyValuePairs)
+        {
+            return new Metadata(keyValuePairs);
+        }
+
+        public static IMetadata With(IDictionary<string, string> keyValuePairs)
+        {
+            return new Metadata(keyValuePairs);
+        }
+
         public ISourceId SourceId
         {
             get { return GetMetadataValue(MetadataKeys.SourceId, v => new SourceId(v)); }
@@ -97,21 +116,82 @@ namespace EventFlow.Aggregates
             set { Add(MetadataKeys.AggregateName, value); }
         }
 
-        public EventMetadata() { }
+        public Metadata() { }
 
-        public EventMetadata(IDictionary<string, string> keyValuePairs)
+        public Metadata(IDictionary<string, string> keyValuePairs)
             : base(keyValuePairs)
         {
         }
 
-        public EventMetadata(IEnumerable<KeyValuePair<string, string>> keyValuePairs)
-            : base(keyValuePairs)
+        public Metadata(IEnumerable<KeyValuePair<string, string>> keyValuePairs)
+            : base(keyValuePairs.ToDictionary(kv => kv.Key, kv => kv.Value))
         {
         }
 
-        public EventMetadata(params KeyValuePair<string, string>[] keyValuePairs)
-            : base(keyValuePairs)
+        public Metadata(params KeyValuePair<string, string>[] keyValuePairs)
+            : this((IEnumerable<KeyValuePair<string, string>>) keyValuePairs)
         {
+        }
+
+        public IMetadata CloneWith(params KeyValuePair<string, string>[] keyValuePairs)
+        {
+            return CloneWith((IEnumerable<KeyValuePair<string, string>>) keyValuePairs);
+        }
+
+        public IMetadata CloneWith(IEnumerable<KeyValuePair<string, string>> keyValuePairs)
+        {
+            var metadata = new Metadata(this);
+            foreach (var kv in keyValuePairs)
+            {
+                if (metadata.ContainsKey(kv.Key))
+                {
+                    throw new ArgumentException($"Key '{kv.Key}' is already present!");
+                }
+                metadata[kv.Key] = kv.Value;
+            }
+            return metadata;
+        }
+
+        public void AddRange(params KeyValuePair<string, string>[] keyValuePairs)
+        {
+            AddRange((IEnumerable<KeyValuePair<string, string>>) keyValuePairs);
+        }
+
+        public void AddRange(IEnumerable<KeyValuePair<string, string>> keyValuePairs)
+        {
+            foreach (var keyValuePair in keyValuePairs)
+            {
+                Add(keyValuePair.Key, keyValuePair.Value);
+            }
+        }
+
+        public override string ToString()
+        {
+            return string.Join(Environment.NewLine, this.Select(kv => $"{kv.Key}: {kv.Value}"));
+        }
+
+        public string GetMetadataValue(string key)
+        {
+            return GetMetadataValue(key, s => s);
+        }
+
+        public T GetMetadataValue<T>(string key, Func<string, T> converter)
+        {
+            string value;
+
+            if (!TryGetValue(key, out value))
+            {
+                throw new MetadataKeyNotFoundException(key);
+            }
+
+            try
+            {
+                return converter(value);
+            }
+            catch (Exception e)
+            {
+                throw new MetadataParseException(key, value, e);
+            }
         }
     }
 }
