+++ Source/EventFlow/EventStores/IMetadataProvider.cs
@@ -31,7 +31,7 @@ namespace EventFlow.EventStores
         IEnumerable<KeyValuePair<string, string>> ProvideMetadata<TAggregate, TIdentity>(
             TIdentity id,
             IAggregateEvent aggregateEvent,
-            IEventMetadata eventMetadata)
+            IMetadata metadata)
             where TAggregate : IAggregateRoot<TIdentity>
             where TIdentity : IIdentity;
     }
