+++ Source/EventFlow/EventStores/SerializedEvent.cs
@@ -29,18 +29,18 @@ namespace EventFlow.EventStores
         public string SerializedMetadata { get; }
         public string SerializedData { get; }
         public int AggregateSequenceNumber { get; }
-        public IEventMetadata Metadata { get; }
+        public IMetadata Metadata { get; }
 
         public SerializedEvent(
             string serializedMetadata,
             string serializedData,
             int aggregateSequenceNumber,
-            IEventMetadata eventMetadata)
+            IMetadata metadata)
         {
             SerializedMetadata = serializedMetadata;
             SerializedData = serializedData;
             AggregateSequenceNumber = aggregateSequenceNumber;
-            Metadata = eventMetadata;
+            Metadata = metadata;
         }
     }
 }
